backend/iProc/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20200713; python_version >= '2.7'
django-cors-headers==3.7.0
django-import-export==2.5.0
django==3.1.7
djangorestframework-jwt==1.11.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
markuppy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pillow==8.1.2
psycopg2==2.8.6
pyjwt==1.7.1
pytz==2021.1
pyyaml==5.4.1
sqlparse==0.4.1; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.0.0; python_version >= '3.6'
xlrd==2.0.1
xlwt==1.3.0
